+++ specifyweb/frontend/js_src/lib/components/querybuilderfieldfilter.tsx
@@ -10,7 +10,7 @@ import type { QueryField } from '../querybuilderutils';
 import { schema } from '../schema';
 import type { RA, RR } from '../types';
 import { defined } from '../types';
-import type { InvalidParseResult, Parser, ValidParseResult } from '../uiparse';
+import type { InvalidParseResult, Parser } from '../uiparse';
 import {
   getValidationAttributes,
   parseValue,
@@ -54,6 +54,8 @@ export const filtersWithDefaultValue: Set<QueryFieldFilter> = new Set([
   'equal',
   'in',
 ]);
+
+const strict = false;
 
 function QueryInputField({
   currentValue,
@@ -127,7 +129,8 @@ function QueryInputField({
     }: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>): void => {
       const input = target as HTMLInputElement;
 
-      if (hasNativeErrors(input)) return;
+      const hasErrors = hasNativeErrors(input);
+      if (strict && hasErrors) return;
 
       const parseResults = extractValues(target)
         .map(f.trim)
@@ -136,12 +139,21 @@ function QueryInputField({
       const errorMessages = parseResults
         .filter((result): result is InvalidParseResult => !result.isValid)
         .map(({ reason, value }) => `${reason} (${value})`);
-      if (errorMessages.length > 0) {
+
+      // Don't display validation messages if native errors are already shown
+      if (errorMessages.length > 0 && !hasErrors) {
         setValidation(errorMessages);
-        return;
+        if (strict) return;
       }
 
-      const validResults = parseResults as RA<ValidParseResult>;
+      const validResults = parseResults.map((result) =>
+        result.isValid
+          ? result
+          : {
+              ...result,
+              parsed: result.value,
+            }
+      );
       const parsed = validResults.some(
         ({ parsed }) => typeof parsed === 'object'
       )